chore: :wrench: update package metadata (#3)
* chore: :wrench: add repository information to package.json

Included the repository details for better project visibility and management.

* chore: :wrench: add author information to package.json

Included author details for better repository attribution.

* chore: :wrench: add license field to package.json

This update includes the addition of the license field to specify the project's licensing information.

* feat: :sparkles: add description field to package.json

Enhances package.json by including a description for better clarity on the project's purpose.

* chore: :rocket: bump version to 0.2.0-alpha.2

Update version in package.json and package-lock.json to reflect the new alpha release.